Tyrannosaurus Animations, Integer ages and fix male Tyrannosaurus texture

+++ src/main/java/net/timeless/jurassicraft/dinosaur/Dinosaur.java
@@ -35,7 +35,7 @@ public abstract class Dinosaur
 
     public abstract double getAdultKnockback();
 
-    public abstract double getMaximumAge();
+    public abstract int getMaximumAge();
 
     public abstract String[] getMaleTextures();
